LanguageClient: Fix possible crash on shutdown
If a shutdown is in progress we need an immediately delete
of the client to avoid accessing an already deleted
TextMarkRegistry instance when deleteLater() is triggered
too late.

+++ src/plugins/languageclient/languageclientmanager.cpp
@@ -125,7 +125,10 @@ void LanguageClientManager::deleteClient(Client *client)
     QTC_ASSERT(client, return);
     client->disconnect();
     managerInstance->m_clients.removeAll(client);
-    client->deleteLater();
+    if (managerInstance->m_shuttingDown)
+        delete client;
+    else
+        client->deleteLater();
 }
 
 void LanguageClientManager::shutdown()
